data types again

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -6,7 +6,7 @@
   * Return: count
   */
 
-unsigned int count(int x)
+unsigned int count(unsigned long int x)
 {
 	unsigned int count;
